I have changed couple of things:
The logics of movement for the tanks has been put under handler class instead of keyboardinput class for better structure of coding. The logics has been adjusted on turning thus now the two tanks can be moved by keyboard perfectly (perfectly meet the class rule- 16 directions of turns, 22.5 degrees each). Yesssss! I have a bit of changes on the game loop for better controlling on the FPS, yet it is not really working. The game works fine yet on on the desire FPS rate. Ben