Hi Jack,
I have loaded the tank images for both players without affecting the tank control. I have also loaded the bullet and powerup image in the texture class. I will also push the "img" folder for your image source for the project. Remeber the files inside the "main" folder should be put under "src" folder and the "img" folder should be put in the as directory as the ".project" file (created by eclipse), and refresh it on eclipse. As the game cannot be run if the directory is not the correct. Ben