Hi Jack,
Finally I got the State Manager Working!! (Running without errors!!) Have been spent for an entire day to find the bug!!! It is actually the initiallization for texture and handler object. Beware of not removing both of them in the initialize method in game class. Cos it will screw up the entire program!. Cheers!

Ben